DDCNL-9071: Removed BBSI and Moved npsIabdSummary
