+++ test/uk/gov/hmrc/tai/model/domain/formatters/BaseTaxAccountHodFormattersSpec.scala
@@ -19,10 +19,12 @@ package uk.gov.hmrc.tai.model.domain.formatters
 import org.scalatestplus.play.PlaySpec
 import play.api.libs.json.{JsArray, JsObject, Json}
 import uk.gov.hmrc.domain.{Generator, Nino}
+import uk.gov.hmrc.tai.model.domain.NpsIabdSummary
+import uk.gov.hmrc.tai.model.domain.NpsIabdSummary.iabdsFromTotalLiabilityReads
 
 import scala.util.Random
 
-class BaseTaxAccountHodFormattersSpec extends PlaySpec with BaseTaxAccountHodFormatters {
+class BaseTaxAccountHodFormattersSpec extends PlaySpec {
 
   "iabdsFromTotalLiabilityReads" must {
